Fix texture name case

+++ utils/image_convert.py
@@ -6,6 +6,7 @@
 # Version included in this tool: January 9, 2021 Release:
 # https://github.com/microsoft/DirectXTex/releases/tag/jan2021
 import math
+import os
 import os.path as op
 import shutil
 import subprocess
@@ -20,7 +21,13 @@ TEXCONV_PATH = op.join(op.dirname(__file__), 'DirectXTex', 'texconv.exe')
 CONVERT_COMAMND = [TEXCONV_PATH, '-y', '-m', '13', '-f']
 
 
-def convert_image(fpath: str, out_fpath: str, type_: str, size: tuple):
+def upper_name_exists(fpath: str) -> bool:
+    """ Determine if the upper-case version of the provided fpath exists. """
+    dir_, fname = op.split(fpath)
+    return fname.upper() in os.listdir(dir_)
+
+
+def convert_image(fpath: str, out_fpath: str, type_: str, size: tuple) -> str:
     """ Take an image and convert it to .DDS. """
     # We can only generate as many mip maps as the image will allow based on
     # size.
@@ -33,11 +40,24 @@ def convert_image(fpath: str, out_fpath: str, type_: str, size: tuple):
     command = CONVERT_COMAMND + [FORMATS[type_], fpath, '-o',
                                  op.dirname(out_fpath)]
     subprocess.call(command)
-    if not op.exists(out_fpath):
+    if not upper_name_exists(out_fpath):
         # In this case the actual filename may be different. Move the converted
         # file to have this new file name.
-        shutil.move(
-            op.join(op.dirname(out_fpath),
-                    op.splitext(op.basename(fpath))[0]) + '.DDS',
-            out_fpath)
+        # If the file is deemed to exist by windows then moving it to have a
+        # different case will fail to do anything. We instead need to rename it
+        # then rename it back with the changed case.
+        if not op.exists(out_fpath):
+            # In this case the file doesn't exist at all, so we can do the
+            # simple move.
+            shutil.move(
+                op.join(op.dirname(out_fpath),
+                        op.splitext(op.basename(fpath))[0]).upper() + '.DDS',
+                out_fpath)
+        else:
+            # Rename to some dummy name then back.
+            shutil.move(
+                op.join(op.dirname(out_fpath),
+                        op.splitext(op.basename(fpath))[0]).upper() + '.DDS',
+                out_fpath + '.DUMMY')
+            shutil.move(out_fpath + '.DUMMY', out_fpath)
     return out_fpath
